+++ NEWS.md
@@ -5,6 +5,7 @@
 * In Target Markdown, add a new `tar_interactive` global option to select interactive mode or non-interactive mode (#469).
 * Highlight a graph neighborhood when the user clicks a node. Control the neighborhood degree with new arguments `degree_from` and `degree_to` of `tar_visnetwork()` and `tar_glimpse()` (#474, @rgayler).
 * Make the target script path configurable in `tar_config_set()` (#476).
+* Add a `tar_script` chunk option in Target Markdown to control where the `{targets}` language engine writes the target script and helper scripts (#478).
 
 ## Deprecations